fix(benchmark): close binary and structure review gaps
